refactor: Use new components on drive view
